+++ sdk/instruction/Cargo.toml
@@ -18,8 +18,6 @@ serde_derive = { workspace = true, optional = true }
 solana-frozen-abi = { workspace = true, optional = true }
 solana-frozen-abi-macro = { workspace = true, optional = true }
 solana-pubkey = { workspace = true, default-features = false }
-solana-sanitize = { workspace = true }
-solana-short-vec = { workspace = true, optional = true }
 
 [target.'cfg(target_os = "solana")'.dependencies]
 solana-define-syscall = { workspace = true }
@@ -46,7 +44,6 @@ frozen-abi = [
 serde = [
     "dep:serde",
     "dep:serde_derive",
-    "dep:solana-short-vec",
     "solana-pubkey/serde",
 ]
