+++ sdk/instruction/src/lib.rs
@@ -15,14 +15,11 @@
 
 #[cfg(feature = "borsh")]
 use borsh::BorshSerialize;
+#[cfg(feature = "serde")]
+use serde_derive::{Deserialize, Serialize};
+use solana_pubkey::Pubkey;
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::wasm_bindgen;
-#[cfg(feature = "serde")]
-use {
-    serde_derive::{Deserialize, Serialize},
-    solana_short_vec as short_vec,
-};
-use {solana_pubkey::Pubkey, solana_sanitize::Sanitize};
 pub mod account_meta;
 pub use account_meta::AccountMeta;
 pub mod error;
@@ -274,57 +271,6 @@ impl Instruction {
             accounts,
             data: data.to_vec(),
         }
-    }
-}
-
-/// A compact encoding of an instruction.
-///
-/// A `CompiledInstruction` is a component of a multi-instruction [`Message`],
-/// which is the core of a Solana transaction. It is created during the
-/// construction of `Message`. Most users will not interact with it directly.
-///
-/// [`Message`]: https://docs.rs/solana-program/latest/solana_program/message/legacy/struct.Message.html
-#[cfg_attr(feature = "frozen-abi", derive(solana_frozen_abi_macro::AbiExample))]
-#[cfg_attr(
-    feature = "serde",
-    derive(Serialize, Deserialize),
-    serde(rename_all = "camelCase")
-)]
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct CompiledInstruction {
-    /// Index into the transaction keys array indicating the program account that executes this instruction.
-    pub program_id_index: u8,
-    /// Ordered indices into the transaction keys array indicating which accounts to pass to the program.
-    #[cfg_attr(feature = "serde", serde(with = "short_vec"))]
-    pub accounts: Vec<u8>,
-    /// The program input data.
-    #[cfg_attr(feature = "serde", serde(with = "short_vec"))]
-    pub data: Vec<u8>,
-}
-
-impl Sanitize for CompiledInstruction {}
-
-impl CompiledInstruction {
-    #[cfg(feature = "bincode")]
-    pub fn new<T: serde::Serialize>(program_ids_index: u8, data: &T, accounts: Vec<u8>) -> Self {
-        let data = bincode::serialize(data).unwrap();
-        Self {
-            program_id_index: program_ids_index,
-            accounts,
-            data,
-        }
-    }
-
-    pub fn new_from_raw_parts(program_id_index: u8, data: Vec<u8>, accounts: Vec<u8>) -> Self {
-        Self {
-            program_id_index,
-            accounts,
-            data,
-        }
-    }
-
-    pub fn program_id<'a>(&self, program_ids: &'a [Pubkey]) -> &'a Pubkey {
-        &program_ids[self.program_id_index as usize]
     }
 }
